# vmm: prefault memory in parallel to optimize boot time