On guests with large amounts of memory, using the `prefault` option can lead to a very long boot time. This PR implements the strategy taken by QEMU to prefault memory in parallel using multiple threads which was first introduced by this [RFC](https://patchwork.ozlabs.org/project/qemu-devel/patch/1483601042-6435-1-git-send-email-jitendra.kolhe@hpe.com/), decreasing time the allocate memory for large guest by an order of magnitude or more.

The implementation generally follows that of [`qemu_prealloc_mem`](https://github.com/qemu/qemu/blob/bd2e12310b18b51aefbf834e6d54989fd175976f/util/oslib-posix.c#L500), simplifying where possible. The essence is that, rather than prefaulting memory by calling `mmap` with the `MAP_POPULATE` flag, this implementation prefaults memory by calling `madvise` with the `MADV_POPULATE_WRITE` flag after it has already been `mmap`ed. This (1) contains the multithreading logic to the `prefault` case because `mmap` without `MAP_POPULATE` is already fast enough without multithreading and (2) allows the prefaulting to be done after applying the NUMA policy because it is much cheaper to apply the NUMA policy to `mmap`ed but not-yet-prefaulted memory.

This implementation also includes the optimization included in this [corresponding QEMU patch](https://patchwork.kernel.org/project/qemu-devel/patch/1582632454-16491-30-git-send-email-pbonzini@redhat.com/) to wait until all threads have been spawned before calling `madvise` on their respective sections to avoid contention over `mmap_sem`, which was found to prevent subsequent threads from spawning.

Below are some stats for a 704 GiB guest with 1 NUMA node configured to use 1GiB hugepages:

## Single-Threaded (existing code on `main`)

| Guest Memory | NUMA Nodes | Hugepage Size | Memory Allocation Time |
| ------------ | ---------: | ------------: | ---------------------: |
| 704 GiB      |          1 |         1 GiB |           81.44134669s |

```
...
cloud-hypervisor: 14.454699ms: <vmm> INFO:vmm/src/api/mod.rs:679 -- API request event: VmBoot
cloud-hypervisor: 16.880310ms: <vmm> INFO:vmm/src/memory_manager.rs:584 -- create ram region for zone mem0, region_start: 0x0, region_size: 0xc0000000
cloud-hypervisor: 367.294435ms: <vmm> INFO:vmm/src/memory_manager.rs:584 -- create ram region for zone mem0, region_start: 0x100000000, region_size: 0xaf40000000
cloud-hypervisor: 81.458227s: <vmm> INFO:vmm/src/vm.rs:501 -- Booting VM from config: ... memory: MemoryConfig { size: 0, mergeable: false, hotplug_method: Acpi, hotplug_size: None, hotplugged_size: None, shared: false, hugepages: false, hugepage_size: None, prefault: false, zones: Some([MemoryZoneConfig { id: "mem0", size: 755914244096, file: None, shared: false, hugepages: true, hugepage_size: Some(1073741824), host_numa_node: Some(0), hotplug_size: None, hotplugged_size: None, prefault: true }]), thp: true } ...
...
```

## Multi-Threaded with 16 threads

| Guest Memory | NUMA Nodes | Hugepage Size | Memory Allocation Time |
| ------------ | ---------: | ------------: | ---------------------: |
| 704 GiB      |          1 |         1 GiB |           6.865287881s |

```
...
cloud-hypervisor: 15.034819ms: <vmm> INFO:vmm/src/api/mod.rs:723 -- API request event: VmBoot
cloud-hypervisor: 17.474119ms: <vmm> INFO:vmm/src/memory_manager.rs:587 -- create ram region for zone mem0, region_start: 0x0, region_size: 0xc0000000
cloud-hypervisor: 214.280206ms: <vmm> INFO:vmm/src/memory_manager.rs:587 -- create ram region for zone mem0, region_start: 0x100000000, region_size: 0xaf40000000
cloud-hypervisor: 6.882762s: <vmm> INFO:vmm/src/vm.rs:502 -- Booting VM from config: memory: MemoryConfig { size: 0, mergeable: false, hotplug_method: Acpi, hotplug_size: None, hotplugged_size: None, shared: false, hugepages: false, hugepage_size: None, prefault: false, zones: Some([MemoryZoneConfig { id: "mem0", size: 755914244096, file: None, shared: false, hugepages: true, hugepage_size: Some(1073741824), host_numa_node: Some(0), hotplug_size: None, hotplugged_size: None, prefault: true }]), thp: true }
...
```

## References

- [QEMU: \[RFC\] mem-prealloc: Reduce large guest start-up and migration time.](https://patchwork.ozlabs.org/project/qemu-devel/patch/1483601042-6435-1-git-send-email-jitendra.kolhe@hpe.com/)
- [QEMU mem-prealloc: optimize large guest startup](https://patchwork.kernel.org/project/qemu-devel/patch/1582632454-16491-30-git-send-email-pbonzini@redhat.com/)